+++ src/main/java/nordmods/iobvariantloader/mixin/client/isleofberk/dragon/BaseRendererMixin.java
@@ -4,19 +4,8 @@ import com.GACMD.isleofberk.entity.base.dragon.ADragonBase;
 import com.GACMD.isleofberk.entity.base.render.render.BaseRenderer;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.mojang.math.Vector3f;
-import net.minecraft.CrashReport;
-import net.minecraft.ReportedException;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.player.LocalPlayer;
-import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
-import net.minecraft.world.entity.Entity;
-import nordmods.iobvariantloader.IoBVariantLoaderClient;
-import nordmods.iobvariantloader.util.ducks.ModelSizeProvider;
-import nordmods.iobvariantloader.util.ducks.VLGlowLayerHelper;
+import nordmods.iobvariantloader.util.layer.PassengerLayer;
 import nordmods.iobvariantloader.util.layer.VLGlowLayer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -24,16 +13,12 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import software.bernie.geckolib3.core.IAnimatable;
 import software.bernie.geckolib3.geo.render.built.GeoBone;
-import software.bernie.geckolib3.geo.render.built.GeoModel;
 import software.bernie.geckolib3.model.AnimatedGeoModel;
 import software.bernie.geckolib3.renderers.geo.GeoEntityRenderer;
-import software.bernie.geckolib3.util.RenderUtils;
-
-import javax.annotation.Nullable;
-import java.util.List;
+import software.bernie.geckolib3.util.EModelRenderCycle;
 
 @Mixin(BaseRenderer.class)
-public abstract class BaseRendererMixin<T extends ADragonBase & IAnimatable> extends GeoEntityRenderer<T> implements VLGlowLayerHelper<T> {
+public abstract class BaseRendererMixin<T extends ADragonBase & IAnimatable> extends GeoEntityRenderer<T>{
     public BaseRendererMixin(EntityRendererProvider.Context renderManager, AnimatedGeoModel<T> modelProvider) {
         super(renderManager, modelProvider);
     }
@@ -46,52 +31,10 @@ public abstract class BaseRendererMixin<T extends ADragonBase & IAnimatable> ext
     }
 
     @Override
-    public void reRender(GeoModel model, T animatable, float partialTicks, RenderType type, PoseStack matrixStackIn,
-                         @Nullable MultiBufferSource renderTypeBuffer, @Nullable VertexConsumer vertexBuilder, int packedLightIn,
-                         int packedOverlayIn, float red, float green, float blue, float alpha) {
-        super.render(model, animatable, partialTicks, type, matrixStackIn, renderTypeBuffer, vertexBuilder, packedLightIn, packedOverlayIn, red, green, blue, alpha);
-    }
-
-    @Override
     public void renderRecursively(GeoBone bone, PoseStack stack, VertexConsumer bufferIn, int packedLightIn,
                                   int packedOverlayIn, float red, float green, float blue, float alpha) {
-        List<Entity> passengers = animatable.getPassengers();
-        for (int i = 0; i < passengers.size(); i++) {
-            Entity passenger = passengers.get(i);
-            String name = "passenger" + i;
-            if (modelProvider.getAnimationProcessor().getBone(name) == null) {
-                IoBVariantLoaderClient.PASSENGERS.remove(passenger.getUUID());
-                continue;
-            }
-            if (bone.name.equals(name)) {
-                IoBVariantLoaderClient.PASSENGERS.remove(passenger.getUUID());
-                float deltaFrameTime = Minecraft.getInstance().getDeltaFrameTime();
-                stack.pushPose();
-                float scale = 1/animatable.getScale() / ((ModelSizeProvider)animatable).getModelSize();
-
-                stack.translate(0, (passenger.getMyRidingOffset() * 2) * scale, 0);
-                RenderUtils.moveToPivot(bone, stack);
-                stack.mulPose(Vector3f.YN.rotationDegrees(180f - passenger.getViewYRot(deltaFrameTime)));
-                stack.scale(scale, scale, scale);
-                renderPassenger(passenger, deltaFrameTime, stack, getCurrentRTB(), packedLightIn);
-                bufferIn = getCurrentRTB().getBuffer(getRenderType(animatable, deltaFrameTime, stack, getCurrentRTB(), bufferIn, packedLightIn, getTextureLocation(animatable)));
-                stack.popPose();
-                IoBVariantLoaderClient.PASSENGERS.add(passenger.getUUID());
-            }
-        }
+        if (getCurrentModelRenderCycle() == EModelRenderCycle.INITIAL)
+            PassengerLayer.renderPassenger(animatable, modelProvider, bone, stack, bufferIn, this, packedLightIn);
         super.renderRecursively(bone, stack, bufferIn, packedLightIn, packedOverlayIn, red, green, blue, alpha);
-    }
-
-    private <E extends Entity> void renderPassenger(E entityIn, float partialTicks, PoseStack matrixStack, MultiBufferSource bufferIn, int packedLight) {
-        boolean isFirstPerson = Minecraft.getInstance().options.getCameraType().isFirstPerson();
-        LocalPlayer clientPlayer = Minecraft.getInstance().player;
-        if (isFirstPerson && entityIn == clientPlayer) return;
-
-        EntityRenderer<? super E> render = Minecraft.getInstance().getEntityRenderDispatcher().getRenderer(entityIn);
-        try {
-            render.render(entityIn, 0, partialTicks, matrixStack, bufferIn, packedLight);
-        } catch (Throwable throwable1) {
-            throw new ReportedException(CrashReport.forThrowable(throwable1, "Rendering entity in world"));
-        }
     }
 }