+++ src/main/java/nordmods/iobvariantloader/mixin/client/isleofberk/egg/ADragonLargeEggRendererMixin.java
@@ -8,7 +8,6 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
-import nordmods.iobvariantloader.util.ducks.VLGlowLayerHelper;
 import nordmods.iobvariantloader.util.layer.VLGlowLayer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -19,7 +18,7 @@ import software.bernie.geckolib3.renderers.geo.GeoProjectilesRenderer;
 import javax.annotation.Nullable;
 
 @Mixin(ADragonLargeEggRenderer.class)
-public abstract class ADragonLargeEggRendererMixin extends GeoProjectilesRenderer<ADragonLargeEggBase> implements VLGlowLayerHelper<ADragonLargeEggBase> {
+public abstract class ADragonLargeEggRendererMixin extends GeoProjectilesRenderer<ADragonLargeEggBase>{
     @Unique
     private final VLGlowLayer<ADragonLargeEggBase> glowLayer = new VLGlowLayer<>(this);
 
@@ -34,13 +33,6 @@ public abstract class ADragonLargeEggRendererMixin extends GeoProjectilesRendere
                        int packedOverlayIn, float red, float green, float blue, float alpha) {
         super.render(model, animatable, partialTicks, type, matrixStackIn, renderTypeBuffer, vertexBuilder, packedLightIn, packedOverlayIn, red, green, blue, alpha);
         glowLayer.render(matrixStackIn, renderTypeBuffer, packedLightIn, animatable, 0, 0, partialTicks, animatable.getAge(), 0, 0);
-    }
-
-    @Override
-    public void reRender(GeoModel model, ADragonLargeEggBase animatable, float partialTicks, RenderType type, PoseStack matrixStackIn,
-                  @Nullable MultiBufferSource renderTypeBuffer, @Nullable VertexConsumer vertexBuilder, int packedLightIn,
-                  int packedOverlayIn, float red, float green, float blue, float alpha) {
-        super.render(model, animatable, partialTicks, type, matrixStackIn, renderTypeBuffer, vertexBuilder, packedLightIn, packedOverlayIn, red, green, blue, alpha);
     }
 
     //*screams*