+++ src/main/java/nordmods/iobvariantloader/mixin/client/isleofberk/egg/MediumEggRendererMixin.java
@@ -8,7 +8,6 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
-import nordmods.iobvariantloader.util.ducks.VLGlowLayerHelper;
 import nordmods.iobvariantloader.util.layer.VLGlowLayer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -19,7 +18,7 @@ import software.bernie.geckolib3.renderers.geo.GeoProjectilesRenderer;
 import javax.annotation.Nullable;
 
 @Mixin(MediumEggRenderer.class)
-public abstract class MediumEggRendererMixin extends GeoProjectilesRenderer<ADragonMediumEggBase> implements VLGlowLayerHelper<ADragonMediumEggBase> {
+public abstract class MediumEggRendererMixin extends GeoProjectilesRenderer<ADragonMediumEggBase>{
     @Unique
     private final VLGlowLayer<ADragonMediumEggBase> glowLayer = new VLGlowLayer<>(this);
 
@@ -33,13 +32,6 @@ public abstract class MediumEggRendererMixin extends GeoProjectilesRenderer<ADra
                        int packedOverlayIn, float red, float green, float blue, float alpha) {
         super.render(model, animatable, partialTicks, type, matrixStackIn, renderTypeBuffer, vertexBuilder, packedLightIn, packedOverlayIn, red, green, blue, alpha);
         glowLayer.render(matrixStackIn, renderTypeBuffer, packedLightIn, animatable, 0, 0, partialTicks, animatable.getAge(), 0, 0);
-    }
-
-    @Override
-    public void reRender(GeoModel model, ADragonMediumEggBase animatable, float partialTicks, RenderType type, PoseStack matrixStackIn,
-                  @Nullable MultiBufferSource renderTypeBuffer, @Nullable VertexConsumer vertexBuilder, int packedLightIn,
-                  int packedOverlayIn, float red, float green, float blue, float alpha) {
-        super.render(model, animatable, partialTicks, type, matrixStackIn, renderTypeBuffer, vertexBuilder, packedLightIn, packedOverlayIn, red, green, blue, alpha);
     }
 
     //*screams*